fix: move tests
